requirements.txt
Pillow>=10.0.0,<11.0.0
python-dotenv>=1.0.0,<2.0.0
fastapi>=0.104.0,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0
python-multipart>=0.0.6,<1.0.0
